+++ R/extract_datim.R
@@ -19,6 +19,11 @@ extract_datim <- function(url,username,password) {
 
   check_internet()
 
+  package_check('httr')
+  package_check('jsonlite')
+  package_check('plyr')
+  package_check('tibble')
+
   if(missing(username))
     username <- datim_user()