+++ R/utilities.R
@@ -5,6 +5,8 @@
 
 check_internet <- function(){
 
+  package_check('curl')
+
   if(!curl::has_internet())
     stop("No internet connection. Cannot excecute API.")
 }
@@ -18,6 +20,8 @@ check_internet <- function(){
 #' @keywords internal
 
 check_status <- function(res){
+
+  package_check('httr')
 
   if(httr::status_code(res) == 200)
     stop("The API returned an error")
@@ -46,6 +50,8 @@ package_check <- function(pkg){
 #' @keywords internal
 
 is_stored <- function(service = c("datim", "email", "s3")){
+
+  package_check('keyring')
 
   accounts <- keyring::key_list()$service
 
@@ -112,3 +118,39 @@ connect_text <- function(txt,
 #' @export
 #' @description negate `%in%`
 `%ni%` <- Negate(`%in%`)
+
+
+#' Generate Temporary Folder
+#'
+#' `temp_folder` created a temporary folder in your AppData directory, which
+#' will be automatically removed after you close your RStudio session.
+#'
+#' @param launch do you want to launch the temp folder in the Windows Explorer?
+#'  default = FALSE
+#'
+#' @return creates a temp directory and stores it as `folderpath_tmp`
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' load_secrets()
+#' temp_folder(launch = TRUE)
+#' purrr::walk2(.x = df_googlefiles$id,
+#'              .y = df_googlefiles$filename,
+#'              .f = ~googledrive::drive_download(googledrive::as_id(.x),
+#'                                                file.path(folderpath_tmp, .y))) }
+temp_folder <- function(launch = FALSE){
+
+  package_check('fs')
+
+  folderpath_tmp <<- fs::dir_create(fs::file_temp())
+
+  usethis::ui_info("A temporary folder is now available here: {usethis::ui_path(folderpath_tmp)}")
+  usethis::ui_info("The folder path is stored as the object {usethis::ui_code('folderpath_tmp')}")
+
+
+  if(launch == TRUE)
+    shell.exec(folderpath_tmp)
+
+  invisible(folderpath_tmp)
+}